Sets up `tarpaulin.toml` to test different feature combinations.
