+++ crate/event_loop/src/event_loop/window.rs
@@ -52,6 +52,7 @@ where
     E: Debug,
     UserEvent: Debug,
 {
+    #[cfg_attr(tarpaulin, skip)]
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         let mut debug_struct = f.debug_struct("EventHandler");
 
@@ -77,6 +78,7 @@ where
     /// # Parameters
     ///
     /// * `event_handlers`: The logic to run for each event loop execution.
+    #[cfg_attr(tarpaulin, skip)]
     pub fn new(event_handlers: Vec<EventHandler<E>>) -> Self {
         let winit_event_loop = WinitEventLoop::with_user_event();
 
@@ -93,6 +95,7 @@ where
     /// # Parameters
     ///
     /// * `event_handlers`: The logic to run for each event loop execution.
+    #[cfg_attr(tarpaulin, skip)]
     pub fn new_with_event<UserEvent>(
         event_handlers: Vec<EventHandler<E>>,
     ) -> EventLoop<E, UserEvent> {
@@ -153,6 +156,7 @@ where
     }
 
     /// Returns the `WinitEventLoop`.
+    #[cfg_attr(tarpaulin, skip)]
     pub fn winit_event_loop(&self) -> &WinitEventLoop<UserEvent> {
         &self.winit_event_loop
     }
@@ -165,6 +169,7 @@ where
 {
     type Target = WinitEventLoop<UserEvent>;
 
+    #[cfg_attr(tarpaulin, skip)]
     fn deref(&self) -> &Self::Target {
         &self.winit_event_loop
     }
@@ -175,7 +180,140 @@ where
     E: Error,
     UserEvent: 'static,
 {
+    #[cfg_attr(tarpaulin, skip)]
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.winit_event_loop
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::thread;
+    #[cfg(feature = "rate_limit")]
+    use std::time::Duration;
+
+    use crossbeam_channel::{SendError, Sender};
+
+    #[cfg(feature = "rate_limit")]
+    use crate::RateLimit;
+    use crate::{EventHandler, EventHandlingOutcome, EventLoop};
+
+    #[test]
+    fn run_runs_event_handlers_until_exit_is_signalled() -> Result<(), SendError<()>> {
+        let (done_tx, done_rx) = crossbeam_channel::bounded(1);
+        let (tx, rx) = crossbeam_channel::bounded(10);
+
+        thread::spawn(|| {
+            let event_handler_send = sender(tx);
+            let event_handler_countdown = countdown(3);
+
+            let event_loop =
+                EventLoop::new_any_thread(vec![event_handler_send, event_handler_countdown])
+                    .with_exit_handler(|e| async move {
+                        let _ = done_tx.send(e);
+                    });
+
+            smol::run(event_loop.run());
+        });
+        let run_result = done_rx.recv();
+        assert_eq!(Ok(None), run_result);
+
+        let count = rx.try_iter().collect::<Vec<()>>().len();
+        assert_eq!(3, count);
+
+        Ok(())
+    }
+
+    #[test]
+    fn run_returns_on_first_error() -> Result<(), SendError<()>> {
+        let (done_tx, done_rx) = crossbeam_channel::bounded(1);
+        let (tx, _rx) = crossbeam_channel::bounded(10);
+
+        thread::spawn(|| {
+            let event_handler_send = sender(tx);
+            let event_handler_countdown = countdown(3);
+
+            let event_loop = EventLoop::new_any_thread(vec![
+                event_handler_send,
+                errorer(),
+                event_handler_countdown,
+            ])
+            .with_exit_handler(|e| async move {
+                let _ = done_tx.send(e);
+            });
+
+            smol::run(event_loop.run());
+        });
+
+        let run_result = done_rx.recv();
+        assert_eq!(Ok(Some(SendError(()))), run_result);
+
+        Ok(())
+    }
+
+    #[cfg(feature = "rate_limit")]
+    #[test]
+    fn event_handlers_are_rate_limited_independently() -> Result<(), SendError<()>> {
+        let (done_tx, done_rx) = crossbeam_channel::bounded(1);
+        let (tx0, rx0) = crossbeam_channel::unbounded();
+        let (tx1, rx1) = crossbeam_channel::bounded(10);
+
+        thread::spawn(|| {
+            let event_handler_send_0 = sender(tx0);
+
+            let event_handler_send_1 =
+                sender(tx1).with_rate_limit(RateLimit::interval(Duration::from_millis(2)));
+            let event_handler_countdown =
+                countdown(3).with_rate_limit(RateLimit::interval(Duration::from_millis(3)));
+
+            let event_loop = EventLoop::new_any_thread(vec![
+                event_handler_countdown,
+                event_handler_send_1,
+                event_handler_send_0,
+            ])
+            .with_exit_handler(|e| async move {
+                let _ = done_tx.send(e);
+            });
+            smol::run(event_loop.run());
+        });
+
+        let _ = done_rx.recv();
+
+        let count_0 = rx0.try_iter().collect::<Vec<()>>().len();
+        let count_1 = rx1.try_iter().collect::<Vec<()>>().len();
+
+        assert!(count_0 >= 8, "count_0: {}", count_0);
+        assert!(count_1 >= 2, "count_1: {}", count_1);
+        assert!(count_1 <= 6, "count_1: {}", count_1);
+
+        Ok(())
+    }
+
+    fn sender(tx: Sender<()>) -> EventHandler<SendError<()>> {
+        EventHandler::<SendError<()>>::new(move || {
+            let tx = tx.clone();
+            async move {
+                tx.send(())?;
+
+                Ok(EventHandlingOutcome::Continue)
+            }
+        })
+    }
+
+    fn countdown(mut count: u32) -> EventHandler<SendError<()>> {
+        EventHandler::<SendError<()>>::new(move || {
+            count = count.saturating_sub(1);
+            async move {
+                if count > 0 {
+                    Ok(EventHandlingOutcome::Continue)
+                } else {
+                    Ok(EventHandlingOutcome::Exit)
+                }
+            }
+        })
+    }
+
+    fn errorer() -> EventHandler<SendError<()>> {
+        EventHandler::<SendError<()>>::new(|| async move { Err(SendError(())) })
     }
 }
