--- crate/workspace_tests/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "workspace_tests"
-version = "0.1.0"
-authors = ["Azriel Hoh <azriel91@gmail.com>"]
-edition = "2018"
-
-[lib]
-doctest = false
-
-[dev-dependencies]
-crossbeam = "0.7.3"
-nginee = { path = "../..", default-features = false, features = ["event_loop"] }
-smol = "0.1.4"
-
-[features]
-default = []
-
-rate_limit = ["nginee/rate_limit"]
-window = ["nginee/window"]
