--- crate/workspace_tests/src/lib.rs
@@ -1,13 +1,0 @@
-//! Tests for `nginee`.
-//!
-//! Contains tests for all workspace crates.
-
-#![deny(
-    anonymous_parameters,
-    arithmetic_overflow,
-    array_into_iter,
-    missing_debug_implementations,
-    missing_docs
-)]
-
-mod nginee_event_loop;
