--- crate/workspace_tests/src/nginee_event_loop/error.rs
@@ -1,12 +1,0 @@
-#[cfg(test)]
-mod tests {
-    use nginee::event_loop::Error;
-
-    #[test]
-    fn display_returns_human_readable_message() {
-        assert_eq!(
-            "FPS must be greater than zero.",
-            Error::RateLimitFpsZero.to_string()
-        )
-    }
-}
