--- crate/workspace_tests/src/nginee_event_loop/event_loop.rs
@@ -1,4 +1,0 @@
-#[cfg(not(feature = "window"))]
-mod headless;
-#[cfg(feature = "window")]
-mod window;
