--- crate/workspace_tests/tarpaulin.toml
@@ -1,15 +1,0 @@
-[no_features]
-no-default-features = true
-
-[features_headless]
-features = ["rate_limit"]
-
-[features_windowed]
-features = ["window"]
-
-[features_windowed_rate_limit]
-features = ["rate_limit", "window"]
-
-[report]
-out = ["Html", "Xml"]
-output-dir = "target/tarpaulin"
