+++ tarpaulin.toml
@@ -1,0 +1,23 @@
+[workspace_default]
+workspace = true
+exclude = [
+    "event_loop",
+    "event_loop_console",
+    "event_loop_rate_limit",
+]
+
+[event_loop_rate_limit_on]
+packages = ["nginee_event_loop"]
+features = "rate_limit"
+
+[event_loop_windowed]
+packages = ["nginee_event_loop"]
+features = "window"
+
+[event_loop_windowed_rate_limit]
+packages = ["nginee_event_loop"]
+features = "rate_limit window"
+
+[report]
+out = ["Html", "Xml"]
+output-dir = "target/tarpaulin"
